move convert-data-curr module to CurrencyConverter service
