--- workers/loc.api/sync/helpers/convert-data-curr.js
@@ -1,106 +1,0 @@
-'use strict'
-
-const { getMethodCollMap } = require('../schema')
-
-const _isEmptyStr = (str) => (
-  !str ||
-  typeof str !== 'string'
-)
-
-const _getCandlesSymbFn = (
-  item,
-  {
-    convertTo,
-    symbolFieldName
-  }
-) => (
-  `t${item[symbolFieldName]}${convertTo}`
-)
-
-module.exports = async (dao, data, convSchema) => {
-  const _convSchema = {
-    convertTo: 'USD',
-    symbolFieldName: '',
-    dateFieldName: '',
-    convFields: [{ inputField: '', outputField: '' }],
-    getCandlesSymbFn: _getCandlesSymbFn,
-    ...convSchema
-  }
-  const {
-    convertTo,
-    symbolFieldName,
-    dateFieldName,
-    mts,
-    convFields,
-    getCandlesSymbFn
-  } = _convSchema
-  const candlesSchema = getMethodCollMap().get('_getCandles')
-  const isArr = Array.isArray(data)
-  const elems = isArr
-    ? data
-    : [data]
-  const res = []
-
-  for (const obj of elems) {
-    const isNotObj = !obj || typeof obj !== 'object'
-    const item = isNotObj ? obj : { ...obj }
-
-    res.push(item)
-
-    if (
-      isNotObj ||
-      _isEmptyStr(convertTo) ||
-      _isEmptyStr(symbolFieldName) ||
-      _isEmptyStr(item[symbolFieldName]) ||
-      !Array.isArray(convFields) ||
-      convFields.length === 0
-    ) {
-      continue
-    }
-
-    const candlesSymb = getCandlesSymbFn(item, _convSchema)
-    const end = Number.isInteger(mts)
-      ? mts
-      : item[dateFieldName]
-
-    if (
-      !candlesSymb ||
-      !Number.isInteger(end)
-    ) {
-      continue
-    }
-
-    const candle = await dao.getElemInCollBy(
-      candlesSchema.name,
-      {
-        [candlesSchema.symbolFieldName]: candlesSymb,
-        end,
-        _dateFieldName: [candlesSchema.dateFieldName]
-      },
-      candlesSchema.sort
-    )
-
-    if (
-      !candle ||
-      typeof candle !== 'object' ||
-      !candle.close ||
-      !Number.isFinite(candle.close)
-    ) {
-      continue
-    }
-
-    convFields.forEach(({ inputField, outputField }) => {
-      if (
-        _isEmptyStr(inputField) ||
-        _isEmptyStr(outputField) ||
-        !Number.isFinite(item[inputField])
-      ) {
-        return
-      }
-
-      item[outputField] = item[inputField] * candle.close
-    })
-  }
-
-  return isArr ? res : res[0]
-}
