+++ workers/loc.api/sync/helpers/index.js
@@ -1,6 +1,5 @@
 'use strict'
 
-const convertDataCurr = require('./convert-data-curr')
 const getStartMtsByTimeframe = require('./get-start-mts-by-timeframe')
 const getMtsGroupedByTimeframe = require('./get-mts-grouped-by-timeframe')
 const calcGroupedData = require('./calc-grouped-data')
@@ -13,7 +12,6 @@ const {
 } = require('./utils')
 
 module.exports = {
-  convertDataCurr,
   getStartMtsByTimeframe,
   getMtsGroupedByTimeframe,
   calcGroupedData,
